+++ src/Platforms/CPU/VectorOps.h
