+++ src/QMCHamiltonians/LatticeDeviationEstimator.h
@@ -14,7 +14,7 @@
 
 #include "Particle/ParticleSet.h"
 #include "QMCHamiltonians/OperatorBase.h"
-#include "ParticleBase/ParticleAttribOps.h"
+#include "CPU/VectorOps.h"
 #include "Particle/DistanceTable.h"
 
 namespace qmcplusplus
