+++ src/QMCWaveFunctions/Jastrow/J1OrbitalSoA.h
@@ -20,7 +20,7 @@
 #include "Configuration.h"
 #include <ResourceHandle.h>
 #include "Particle/DistanceTable.h"
-#include "ParticleBase/ParticleAttribOps.h"
+#include "CPU/VectorOps.h"
 #include "QMCWaveFunctions/WaveFunctionComponent.h"
 #include "Utilities/qmc_common.h"
 #include "Utilities/IteratorUtility.h"
