+++ src/QMCWaveFunctions/tests/test_example_he.cpp
@@ -14,7 +14,7 @@
 
 #include "Configuration.h"
 #include "Message/Communicate.h"
-#include "ParticleBase/ParticleAttribOps.h"
+#include "CPU/VectorOps.h"
 #include "OhmmsData/Libxml2Doc.h"
 #include "QMCWaveFunctions/WaveFunctionFactory.h"
 #include "QMCWaveFunctions/ExampleHeComponent.h"
